genai-in-edu/backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
transformers
sentence-transformers
faiss-cpu
langchain
redis
py2neo
motor
celery
pycryptodome
huggingface-hub
python-dotenv
tf-keras
langchain-huggingface
langchain-core
accelerate
neo4j
pillow 
pyttsx3
